whost/pyproject.toml
[tool.poetry]
name = "whost"
version = "0.1.2"
repository = "https://github.com/kiwix/cardshop"
homepage = "https://www.kiwix.org/kiwix-hotspot/"
description = "Cardshop Writer Host Tools"
authors = ["renaud gaudin <reg@kiwix.org>"]
include = ["whost-config", "whost-configurator", "whost-restart-all", "whost-setup", "whost-start-all", "whost-start-worker", "whost-stop-all", "whost-update", "cardshop-host.config.original"]

[tool.poetry.scripts]


[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.21"
PyYaml = "^3.13"
netifaces = "^0.10.7"
colorama = "^0.4.1"
unidecode = "^1.0"
tabulate = "^0.8.2"
ipython = "^7.2"
humanfriendly = "^4.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
